requirements.txt
numpy>=1.24.1
pytest>=7.2.0
pytest-cov>=4.0.0
coverage>=7.2.5, <8.0.0
python-dotenv>=1.0.0
tc_messageBroker>=1.6.3
pymongo>=4.3.3, <5.0.0
neo4j>=5.8.0, <6.0.0
llama-index>=0.9.21, <1.0.0
sqlalchemy[asyncio]
cohere>=4.39, <5.0.0
pgvector
asyncpg
psycopg2-binary
